TV6SydVz: cognito path helpers
Use rails route helpers for fetching the auth path. This way the path
will adjust depending on the environment where the application is
deployed. For example, the route helper will include any base paths that
may exist if the app is deployed to an environment where a base path is
required.